refactor: Move to debug util
